Migrate password login functionality to BasicApiConnection as planned by the deprecation
