+++ wdtk-examples/src/main/java/org/wikidata/wdtk/examples/EditOnlineDataExample.java
@@ -42,6 +42,7 @@ import org.wikidata.wdtk.datamodel.interfaces.StatementDocument;
 import org.wikidata.wdtk.datamodel.interfaces.StatementGroup;
 import org.wikidata.wdtk.util.WebResourceFetcherImpl;
 import org.wikidata.wdtk.wikibaseapi.ApiConnection;
+import org.wikidata.wdtk.wikibaseapi.BasicApiConnection;
 import org.wikidata.wdtk.wikibaseapi.LoginFailedException;
 import org.wikidata.wdtk.wikibaseapi.WikibaseDataEditor;
 import org.wikidata.wdtk.wikibaseapi.WikibaseDataFetcher;
@@ -97,7 +98,7 @@ public class EditOnlineDataExample {
 		WebResourceFetcherImpl
 				.setUserAgent("Wikidata Toolkit EditOnlineDataExample");
 
-		ApiConnection connection = ApiConnection.getTestWikidataApiConnection();
+		ApiConnection connection = BasicApiConnection.getTestWikidataApiConnection();
 		// Optional login -- required for operations on real wikis:
 		// connection.login("my username", "my password");
 		WikibaseDataEditor wbde = new WikibaseDataEditor(connection, siteIri);
